Fixing a bug with git pull step
I omitted this prompt stage from the pull command when I added the new
git checkout command below. This was causing errors because the expect
script plowed through and started executing the git checkout before the
git pull was done. By telling it to wait for the prompt the script will
fully pull the branch before running the checkout.